# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a B2B-API server.",
        "title": "B2B Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8070",
    "basePath": "/api/v1",
    "paths": {
...
        "/2fa/confirm-sms-otp": {
            "get": {
                "description": "SMS Confirmation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2FA"
                ],
                "summary": "SMS Confirmation route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uint",
                        "description": " ",
                        "name": "otp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Token"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/merchant/confirm-transaction": {
            "get": {
                "description": "ConfirmTransaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "ConfirmTransaction route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "json",
                        "description": " ",
                        "name": "transactionId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/merchant/get-account-details": {
            "get": {
                "description": "GetAccountDetails",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "GetAccountDetails route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "json",
                        "description": " ",
                        "name": "accountNumber",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AccountDetailsForBankTrans"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/merchant/get-accounts": {
            "get": {
                "description": "GetAccounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "GetAccounts route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetAccountsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/merchant/get-all-amount": {
            "get": {
                "description": "GetAllAmount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "GetAllAmount route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "json",
                        "description": " ",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllAmountsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/merchant/get-converted-value": {
            "get": {
                "description": "GetConvertedValue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "GetConvertedValue route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "json",
                        "description": " ",
                        "name": "value",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "json",
                        "description": " ",
                        "name": "currencyFrom",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "json",
                        "description": " ",
                        "name": "currencyTo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConvertedValueResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "merchant.AccountDetailsForBankTrans": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "merchant.Accounts": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "currencySym": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "merchant.ConvertedValue": {
            "type": "object",
            "properties": {
                "rate": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "models.AccountDetailsForBankTrans": {
            "type": "object",
            "properties": {
                "local": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/merchant.AccountDetailsForBankTrans"
                    }
                },
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                }
            }
        },
...
        "models.ConvertedValueResponse": {
            "type": "object",
            "properties": {
                "convertedValue": {
                    "$ref": "#/definitions/merchant.ConvertedValue"
                },
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                }
            }
        },
        "models.ErrorModel": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "errorDescription": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "errorDescription": {
                    "type": "string"
                },
                "response": {}
            }
        },
        "models.GetAccountsResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "$ref": "#/definitions/merchant.Accounts"
                },
                "merchName": {
                    "type": "string"
                },
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                }
            }
        },
        "models.GetAllAmountsResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "merchName": {
                    "type": "string"
                },
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                }
            }
        },
...
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                },
                "response": {
                    "$ref": "#/definitions/models.twoFa"
                }
            }
        },
...
        "models.Token": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "models.googleAuth": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                }
            }
        },
        "models.smsOTP": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                }
            }
        },
        "models.twoFa": {
            "type": "object",
            "properties": {
                "googleAuthenticator": {
                    "$ref": "#/definitions/models.googleAuth"
                },
                "smsOtp": {
                    "$ref": "#/definitions/models.smsOTP"
                }
            }
        }
    }
}